Allow navigation to particular tab locations when loading page
This disables a workaround which caused the initial loading to
always end up at the playback page, even if another path was after
the #. A comment claimed that it was a "workaround for lost
history/back view after browser reload", so this might cause
some problems. The workaround is still there, but commented out.